+++ src/test/java/com/chunkr/compress/chunkers/ModifiedChunkerTest.java
@@ -11,7 +11,7 @@ public class ModifiedChunkerTest {
 		boolean[] unchunks = new boolean[] { true, false, true, true, false, true };
 		int[] chunks = new int[] { 5, 3, 6, 5 };
 		
-		Chunker chunker = new ModifiedChunker((byte) 3);
+		Chunker chunker = new ModifiedChunker(3);
 		assertArrayEquals(chunks, chunker.chunk(unchunks));
 	}
 	
@@ -20,7 +20,7 @@ public class ModifiedChunkerTest {
 		boolean[] unchunks = new boolean[] { true, false, true, true, false, true };
 		int[] chunks = new int[] { 5, 3, 6, 5 };
 		
-		Chunker chunker = new ModifiedChunker((byte) 3);
+		Chunker chunker = new ModifiedChunker(3);
 		assertArrayEquals(unchunks, chunker.unchunk(chunks));
 	}
 }
